b9ccd96 Disable Manta plugin in static builds.
4497068 Bring in Python module static build fix from VTK.
44a4061 Exclude Manta warnings.
14139be Merge remote-tracking branch 'stage/pv_warnings' into 13567_fix_manta_view
5b430e5 Adding test for MantaView plugin.
5fe19a3 Fixed warnings.
4da716d Fix issues due to VTK API change.
b9bf414 Fix examples dependency.
d4d1073 Bringing back VTK tests defined in plugins.
723ecde BUG #13567. Fix segfault in MantaView.
5409b7f Update MantaView directory structure to as expected by VTK modular.
7c23c16 Moving Manta to MantaView directory to match the plugin's name.
aefb401 Fix remaining warnings
571efcb h5 Warnings
26eb559 More warning fix

+++ Plugins/StreamingView/VTK/vtkImageNetCDFPOPReader.cxx
@@ -349,13 +349,17 @@ int vtkImageNetCDFPOPReader::RequestData(vtkInformation* request,
   imageData->SetExtent(subext);
 
   //setup extents for netcdf library to read the netcdf data file
-  size_t start[]= {subext[4]*this->Internals->SK,
-                   subext[2]*this->Internals->SJ,
-                   subext[0]*this->Internals->SI};
+  size_t start[]= {
+    static_cast<size_t>(subext[4]*this->Internals->SK),
+    static_cast<size_t>(subext[2]*this->Internals->SJ),
+    static_cast<size_t>(subext[0]*this->Internals->SI)
+  };
 
-  size_t count[]= {subext[5]-subext[4]+1,
-                   subext[3]-subext[2]+1,
-                   subext[1]-subext[0]+1};
+  size_t count[]= {
+    static_cast<size_t>(subext[5]-subext[4]+1),
+    static_cast<size_t>(subext[3]-subext[2]+1),
+    static_cast<size_t>(subext[1]-subext[0]+1)
+  };
 
   ptrdiff_t rStride[3] = { (ptrdiff_t)this->Internals->SK,
                            (ptrdiff_t)this->Internals->SJ,
